refactor(verify-step2): mechanically move verify logic behind facade
